Electron Forge
forge

update

update pkg.json

update pkg.json

update

refactor settings (WIP)

refactor settings (WIP)

refactor settings (WIP)

update

use latest logo

small improvements and fix electron report

attributes

Update .gitattributes

update readme

fix fossa badges
