--- .babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    "es2015",
-    "react",
-    "stage-2"
-    ],
-  "plugins": [
-    "transform-async-to-generator",
-    "transform-react-display-name"
-  ],
-  "sourceMaps": "inline"
-}
