+++ .eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
-    extends: ['aegir', 'standard-react'],
-    parser: 'babel-eslint',
-    parserOptions: {
-        sourceType: 'module'
-    }
+  extends: ['aegir', 'standard-react'],
+  parser: 'babel-eslint',
+  parserOptions: {
+      sourceType: 'module'
+  }
 }
