+++ .gitignore
@@ -5,3 +5,4 @@ release
 npm-debug.log
 app.log
 package-lock.json
+out
