+++ package.json
@@ -1,15 +1,18 @@
 {
   "name": "ipfs-station",
+  "productName": "IPFS Station",
   "version": "0.1.0",
   "description": "IPFS Native Application",
   "main": "src/index.js",
   "dependencies": {
-    "ipfs-api": "^17.1.2",
-    "ipfs-geoip": "^2.1.0",
+    "electron-compile": "^6.4.2",
+    "electron-is-dev": "^0.3.0",
+    "electron-squirrel-startup": "^1.0.0",
+    "ipfs-api": "^17.1.3",
+    "ipfs-geoip": "^2.3.0",
     "ipfs-logo": "github:ipfs/logo",
     "ipfsd-ctl": "^0.25.1",
-    "menubar": "^5.0.0",
-    "minimist": "^1.2.0",
+    "menubar": "^5.2.3",
     "multiaddr": "^3.0.1",
     "node-notifier": "^5.1.2",
     "normalize.css": "^7.0.0",
@@ -20,53 +23,63 @@
     "react-dom": "^16.1.1",
     "react-transition-group": "^2.2.1",
     "react-widgets": "^4.1.1",
-    "winston": "^2.1.0"
+    "winston": "^3.0.0-rc1"
   },
   "devDependencies": {
-    "babel-core": "^6.3.21",
     "babel-eslint": "^8.0.2",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-array-includes": "^2.0.3",
-    "babel-plugin-react-transform": "^3.0.0",
-    "babel-plugin-transform-inline-environment-variables": "^6.3.13",
-    "babel-polyfill": "^6.5.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
-    "caniuse-api": "^2.0.0",
-    "conventional-changelog": "^1.1.0",
-    "conventional-github-releaser": "^2.0.0",
-    "css-loader": "^0.28.7",
-    "devtron": "^1.3.0",
-    "electron-debug": "^1.0.0",
-    "electron-packager": "10.1.0",
-    "electron-prebuilt-compile": "^1.7.9",
-    "eslint": "^4.11.0",
+    "babel-plugin-transform-async-to-generator": "^6.24.1",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "electron-debug": "^1.4.0",
+    "electron-forge": "^4.1.5",
+    "electron-prebuilt-compile": "1.7.9",
+    "eslint": "^4",
     "eslint-config-aegir": "^1.0.1",
-    "eslint-config-standard": "^10.2.1",
     "eslint-config-standard-react": "^5.0.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-react": "^7.5.1",
-    "eslint-plugin-standard": "^3.0.1",
-    "longjohn": "^0.2.11",
-    "pre-commit": "^1.2.1",
-    "react-transform-catch-errors": "^1.0.0",
-    "react-transform-hmr": "^1.0.0",
-    "redbox-react": "^1.0.5",
-    "rimraf": "^2.4.3",
-    "shelljs": "^0.7.0"
+    "eslint-plugin-import": "^2",
+    "eslint-plugin-jsx-a11y": "^6",
+    "eslint-plugin-react": "^7",
+    "pre-commit": "^1.2.2"
   },
   "scripts": {
-    "start": "electron src/index.js",
-    "test": "eslint .",
-    "lint": "eslint .",
-    "dist": "./pkg.js",
-    "dist-all": "./pkg.js --all",
-    "start:prod": "NODE_ENV=production electron src/index.js",
-    "changelog": "conventional-changelog -p angular -i CHANGELOG.md -w",
-    "changelog:github": "conventional-github-releaser -p angular"
+    "start": "electron-forge start",
+    "package": "electron-forge package",
+    "make": "electron-forge make",
+    "publish": "electron-forge publish",
+    "lint": "eslint ."
+  },
+  "config": {
+    "forge": {
+      "make_targets": {
+        "win32": [
+          "squirrel"
+        ],
+        "darwin": [
+          "zip"
+        ],
+        "linux": [
+          "deb",
+          "rpm"
+        ]
+      },
+      "electronPackagerConfig": {
+        "packageManager": "npm"
+      },
+      "electronWinstallerConfig": {
+        "name": "IPFS Station"
+      },
+      "electronInstallerDebian": {},
+      "electronInstallerRedhat": {},
+      "github_repository": {
+        "owner": "",
+        "name": ""
+      },
+      "windowsStoreConfig": {
+        "packageName": "io.ipfs.station",
+        "name": "IPFS Station"
+      }
+    }
   },
   "pre-commit": [
     "lint"
